backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.128.1
greenlet==3.3.1
h11==0.16.0
idna==3.11
psycopg2-binary==2.9.11
pwdlib==0.3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uuid==1.30
uvicorn==0.40.0
